requirements.txt
accelerate==0.26.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aiosqlite==0.22.1
altair==6.0.0
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
azure-core==1.37.0
azure-identity==1.25.1
banks==2.2.0
beautifulsoup4==4.14.3
bitsandbytes==0.49.1
blinker==1.9.0
cachetools==6.2.4
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cohere==5.20.1
colorama==0.4.6
cryptography==46.0.3
Cython==3.2.4
dataclasses-json==0.6.7
decord==0.6.0
defusedxml==0.7.1
Deprecated==1.3.1
dirtyjson==1.0.8
distro==1.9.0
duckdb==1.4.3
eval_type_backport==0.3.1
fastavro==1.12.1
ffmpeg-python==0.2.0
filelock==3.20.2
filetype==1.2.0
frozenlist==1.8.0
fsspec==2026.1.0
future==1.0.0
gitdb==4.0.12
GitPython==3.1.46
googleapis-common-protos==1.72.0
greenlet==3.3.0
griffe==1.15.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==0.36.1
idna==3.11
ImageIO==2.37.2
importlib_metadata==8.7.1
iniconfig==2.3.0
invoke==2.2.1
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
llama-index-core==0.14.12
llama-index-embeddings-adapter==0.4.1
llama-index-embeddings-huggingface==0.6.1
llama-index-experimental==0.6.3
llama-index-finetuning==0.4.1
llama-index-instrumentation==0.4.2
llama-index-llms-azure-openai==0.4.2
llama-index-llms-mistralai==0.7.1
llama-index-llms-ollama==0.9.1
llama-index-llms-openai==0.6.12
llama-index-postprocessor-cohere-rerank==0.5.1
llama-index-readers-file==0.5.6
llama-index-workflows==2.11.7
-e git+https://github.com/haotian-liu/LLaVA.git@c121f0432da27facab705978f83c4ada465e46fd#egg=llava
MarkupSafe==3.0.3
marshmallow==3.26.2
mistralai==1.10.0
mpmath==1.3.0
msal==1.34.0
msal-extensions==1.3.1
multidict==6.7.0
mypy_extensions==1.1.0
narwhals==2.15.0
nest-asyncio==1.6.0
networkx==3.6.1
nltk==3.9.2
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
ollama==0.6.1
openai==2.14.0
opencv-python==4.12.0.88
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
packaging==25.0
pandas==2.2.3
pillow==12.1.0
platformdirs==4.5.1
plotly==6.5.1
pluggy==1.6.0
propcache==0.4.1
protobuf==6.33.2
psutil==7.2.1
pyarrow==22.0.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
PyJWT==2.10.1
pypdf==6.6.0
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rpds-py==0.30.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.16.3
sentence-transformers==3.0.1
sentencepiece==0.2.1
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8.1
SQLAlchemy==2.0.45
streamlit==1.52.2
striprtf==0.0.26
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.21.4
toml==0.10.2
torch==2.9.1
torchvision==0.24.1
tornado==6.5.4
tqdm==4.67.1
transformers==4.50.3
triton==3.5.1
typer-slim==0.21.1
types-requests==2.32.4.20260107
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
watchdog==6.0.0
wrapt==2.0.1
yarl==1.22.0
zipp==3.23.0
